weather data code
I used the data I downloaded from the PRISM website and combined it with the demographic data. I did so only for ppt and tmean. I plan to eventually use the PRISM R package to acquire the data but since its a lot computationally, I am first figuring out who to incorperate the data using the downloaded weather data.